Rename azle_ast dir to azle_act
ACT standing for Abstract Canister Tree. It describes the things that
are in a canister definition/declaration.
