+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/generators/canister_methods/system/init.rs
@@ -1,6 +1,6 @@
 use crate::generators::ic_object::generate_ic_object;
 use crate::{
-    azle_ast::CanisterMethodType,
+    azle_act::CanisterMethodType,
     generators::canister_methods::{
         functions::{
             generate_param_name_idents, generate_param_types, generate_params_token_stream,
