+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/lib.rs
@@ -17,7 +17,7 @@ use swc_ecma_parser::{
 };
 
 use crate::{
-    azle_ast::CanisterMethodType,
+    azle_act::CanisterMethodType,
     generators::{
         azle_into_js_value::generate_azle_into_js_value,
         azle_try_from_js_value::generate_azle_try_from_js_value,
@@ -45,7 +45,7 @@ use crate::{
     },
 };
 
-mod azle_ast;
+mod azle_act;
 mod ts_ast;
 
 pub mod generators;
